requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=9.0.0
python-dotenv>=1.0.0
jazzmin>=3.0.0